fix:refactor registerUser method update createOrganizationbyOwner test
