+++ backend/src/interfaces/organization.service.ts
@@ -1,0 +1,4 @@
+export interface ICreateOrganizationByOwner {
+    organizationSlug: string;
+    name: string;
+  }
